requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.23
psycopg[binary]>=3.1.13
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
ebooklib>=0.18.0
